Increase stack capacity on callsite retrieval
